testing couchdb instance

+++ initialize.py
@@ -1,5 +1,8 @@
 import yaml, json
-
+import requests
+import time
+import subprocess
+import os
 
 def load_config():
     with open("config.json","r") as f:
@@ -29,6 +32,21 @@ def initialize_docker_compose():
                                 'version': '3',
                                 'volumes': {'couchdb-data': None}}
     return docker_compose_yaml
+def wait_for_couch_container():
+    while True:
+        try:
+            if requests.get("http://127.0.0.1:8001/").status_code == 200:
+                print("Successfully got CouchDB container!")
+                return True
+            else:
+                print("Non 200 Response")
+                return False
+        except Exception as e:
+            print(e)
+            t = 3*attempt
+            print("Waiting for container to spin up...{} seconds".format(t))
+            time.sleep(t)
+            attempt += 1
 
 config = load_config()
 docker_compose_yaml = initialize_docker_compose()
@@ -37,11 +55,22 @@ if config['prompt']:
     couch_password = "COUCHDB_PASSWORD={}".format(input("CouchDB Admin Password: "))
     docker_compose_yaml['services']['couchdb']['environment'] = [couch_username, couch_password]
 
-with open("docker-compose.yml","w") as f:
-    f.write(yaml.dump(docker_compose_yaml))
+if 'docker-compose.yml' in os.listdir('.'):
+    answer = input("[?] Are you sure you want to overwrite the existing docker-compose.yml? (Y/n): ")
+    if answer.lower() in ['yes','y']:
+        with open("docker-compose.yml","w") as f:
+            f.write(yaml.dump(docker_compose_yaml))
+    else:
+        print("[-] Skipping writing docker-compose.yml")
 
-# Run docker-compose up (Subprocess)
+print("[+] Running 'docker-compose up' as a child process. If this is the first time this may take a long time...")
+subprocess.Popen(["docker-compose", "up"], shell=True, creationflags=subprocess.CREATE_NEW_CONSOLE)
+attempt = 1
 
+wait_for_couch_container()
+
+print("[+] Destroying containers")
+subprocess.Popen(["docker-compose", "down"], shell=True)
 # Create medblock database
 # create _users database
 # Write Validation documents for medblock
